refactor plotter functions
